pyproject.toml
[project]
name = "ziniao"
version = "0.2.70"
description = "紫鸟与 Chrome 浏览器 AI 自动化 — 让 AI Agent 操控紫鸟店铺与本地 Chrome"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
keywords = ["ziniao", "mcp", "browser", "chrome", "cdp", "e-commerce", "cross-border"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Topic :: Software Development :: Testing",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "requests>=2.28.0",
    "PyYAML>=6.0",
    "mcp>=1.0.0",
    "httpx>=0.27.0",
    "jinja2>=3.1.0",
    # 0.48.1 对 Chrome 147+ 的 CDP schema drift (Cookie.sameParty /
    # privateNetworkRequestPolicy 缺省字段) 会在 _listener 抛 KeyError 并
    # 打印完整 cookie dump；ziniao_mcp.logging_setup 已把 nodriver 日志
    # 压到 ERROR 做兜底，等上游发新版后再上提下限。
    "nodriver>=0.48.1",
    "typer>=0.12.0",
    "rich>=13.0.0",
    # Windows 默认无系统 IANA 时区库；zoneinfo 需 tzdata 才能解析 Asia/Tokyo（乐天日界）。
    "tzdata>=2024.1",
    # Raw CDP WebSocket for input-only automation (no transitive-only reliance).
    "websockets>=12.0",
]

[project.scripts]
ziniao = "ziniao_mcp.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["ziniao", "ziniao_mcp", "ziniao_webdriver"]
artifacts = ["skills/**/SKILL.md", "skills/**/references/*.md", "skills/**/assets/*", "skills/**/scripts/*"]

[tool.hatch.build.targets.wheel.force-include]
"skills" = "skills"

[dependency-groups]
dev = [
    # 可选功能 secret source `keyring:` 的测试与 Pylint 静态解析需要。
    "keyring>=25.0",
    "pylint>=4.0.5",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "python-dotenv>=1.2.2",
    "ruff>=0.9.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
# 项目根下 basetemp，减少 Windows 在 %TEMP% 清理 pytest 目录时的 PermissionError
addopts = "--basetemp=.pytest_basetemp"
markers = [
    "e2e: browser-backed dogfooding tests skipped by default in CI",
]

[tool.ruff]
target-version = "py310"
extend-exclude = ["skills"]

[tool.ruff.lint]
ignore = ["PLC0415", "BLE001"]

[tool.ruff.lint.per-file-ignores]
"tests/integration_test.py" = ["E402"]

# Pylint 与 Ruff 分工：风格/格式以 Ruff 为主；关闭与项目习惯冲突或噪声大的规则。
# 保留「写法更优雅简洁」类检查：use-implicit-booleaness-not-comparison、use-dict-literal、use-maxsplit-arg、reimported 等不设为全局 disable。
# undefined-all-variable：相对导入 + __all__ 在整包扫描时 astroid 误报。
[tool.pylint.main]
py-version = "3.10"
jobs = 0

[tool.pylint."messages control"]
disable = [
    "undefined-all-variable",
    "import-outside-toplevel",
    "broad-exception-caught",
    "duplicate-code",
    "cyclic-import",
    "missing-module-docstring",
    "missing-class-docstring",
    "missing-function-docstring",
    "invalid-name",
    "too-few-public-methods",
    "too-many-lines",
    "too-many-branches",
    "too-many-locals",
    "too-many-statements",
    "too-many-arguments",
    "too-many-positional-arguments",
    "too-many-return-statements",
    "too-many-nested-blocks",
    "too-many-instance-attributes",
    "redefined-outer-name",
    "protected-access",
    "unsubscriptable-object",
    "unused-argument",
    "unused-variable",
    "line-too-long",
    "wrong-import-position",
    "global-statement",
    "too-many-public-methods",
]

# Pylance / pyright：从仓库根目录的 .venv 解析 site-packages（typer 等），
# 避免未选对本机解释器时出现 reportMissingImports。
[tool.pyright]
venvPath = "."
venv = ".venv"
